refactor: coupon utils 역할에따라 디렉토리 구분
